+++ scripts/mixed_3d/pcno_reduce_data.sh
@@ -1,0 +1,15 @@
+#!/bin/bash
+#SBATCH -o logs/PCNO_reduce_data_vertex_centered.out
+#SBATCH --qos=low
+#SBATCH -p C064M1024G
+#SBATCH -J PCNO_reduce_data_vertex_centered
+#SBATCH --nodes=1 
+#SBATCH --ntasks=1
+#SBATCH --cpus-per-task=32
+#SBATCH --time=100:00:00
+
+module load conda
+source activate pytorch
+python pcno_geo_mixed_3d_reduce_data.py --n_train 4000 \
+                                        --n_test 100 \
+                                        --mesh_type "vertex_centered"  # "cell_centered" , "vertex_centered"
